requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.1
Django==4.0.1
django-ninja==0.16.1
django-ninja-extra==0.14.0
django-ninja-jwt==5.1.5
dnspython==2.1.0
email-validator==1.1.3
idna==3.3
injector==0.19.0
ninja-schema==0.12.6
pycparser==2.21
pydantic==1.8.2
PyJWT==2.3.0
sqlparse==0.4.2
typing_extensions==4.0.1
tzdata==2021.5
